% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carrot}
\alias{carrot}
\title{a carrot comparative linkage map data frame
kindly provided by Massimo Iorizzo:
Cavagnaro et al. BMC Genomics 2014, 15:1118}
\format{An object of class \code{data.frame} with 126 rows and 3 columns.}
\usage{
carrot
}
\description{
Contains the following columns:
     \enumerate{
       \item group - This will be the title for the linkage
             group unless overridden.
       \item position - must be in numerical order ascending within
             linkage group name.
       \item locus - marker name at this position.
     }
}
\keyword{datasets}

