% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkageMapView.R
\docType{package}
\name{LinkageMapView}
\alias{LinkageMapView}
\alias{LinkageMapView-package}
\title{LinkageMapView: A package for plotting linkage group maps and QTLs}
\description{
The LinkageMapView package provides one external function: lmv
}
\details{
LinkageMapView produces high resolution, publication ready linkage maps and QTL maps.

There are many optional parameters to format the output pdf.
Please see the help for function lmv for a full description of
each parameter and examples.
}

