% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oat}
\alias{oat}
\title{oat consensus map data frame linkage group Mrg01}
\format{An object of class \code{data.frame} with 16668 rows and 3 columns.}
\usage{
oat
}
\description{
Chaffin, A. S., Y. Huang, S. Smith, W. A. Bekele, E. Babiker, B. N. Gnanesh,
B. J. Foresman, S. G. Blanchard, J. J. Jay, R. W. Reid, C. P. Wight, S. Chao,
R. Oliver, E. Islamovic, F. L. Kolb, C. McCartney, J. W. Mitchell Fetch,
A. D. Beattie, ?. Bjornstad, J. M. Bonman, T. Langdon, C. J. Howarth,
C. R. Brouwer, E. N. Jellen, K. E. Klos, J. A. Poland, T. Hsieh, R. Brown,
E. Jackson, J. A. Schlueter, and N. A. Tinker.
2016. A Consensus Map in Cultivated Hexaploid Oat Reveals Conserved Grass
Synteny with Substantial Subgenome Rearrangement.
Plant Genome 9. doi:10.3835/plantgenome2015.10.0102
}
\details{
Contains the following columns:
     \enumerate{
       \item Group - This will be the title for the linkage
             group unless overridden.
       \item Position - must be in numerical order ascending within
             linkage group name.
       \item Locus - marker name at this position.
     }
}
\keyword{datasets}

