% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinkageMapView.R
\docType{package}
\name{LinkageMapView}
\alias{LinkageMapView}
\title{LinkageMapView: A package for plotting linkage group maps and QTLs}
\description{
LinkageMapView produces high resolution, publication ready linkage maps and QTL maps.
}
\details{
There are many optional parameters to format the output pdf.
Please see the help for function lmv.linkage.plot for a full description of
each parameter and examples.
}
